move tests to onAuth
