+++ __tests__/test-push.js
@@ -1,7 +1,7 @@
 /* eslint-env node, browser, jasmine */
 import http from 'isomorphic-git/http'
 
-const { setConfig, push, listBranches } = require('isomorphic-git')
+const { E, setConfig, push, listBranches } = require('isomorphic-git')
 
 const { makeFixture } = require('./__helpers__/FixtureFS.js')
 
@@ -191,10 +191,9 @@ describe('push', () => {
       fs,
       http,
       gitdir,
-      username: 'testuser',
-      password: 'testpassword',
       remote: 'auth',
       ref: 'master',
+      onAuth: () => ({ username: 'testuser', password: 'testpassword' }),
     })
     expect(res).toBeTruthy()
     expect(res.ok).toBe(true)
@@ -261,14 +260,94 @@ describe('push', () => {
         fs,
         http,
         gitdir,
-        username: 'test',
-        password: 'test',
         remote: 'auth',
         ref: 'master',
+        onAuth: () => ({ username: 'test', password: 'test' }),
       })
     } catch (err) {
       error = err.message
     }
     expect(error).toContain('401')
+  })
+  it('onAuthSuccess', async () => {
+    // Setup
+    const { fs, gitdir } = await makeFixture('test-push')
+    await setConfig({
+      fs,
+      gitdir,
+      path: 'remote.auth.url',
+      value: `http://${localhost}:8888/test-push-server-auth.git`,
+    })
+    // Test
+    let fillCalled = false
+    let approvedCalled = false
+    let rejectedCalled = false
+    await push({
+      fs,
+      http,
+      gitdir,
+      remote: 'auth',
+      ref: 'master',
+      async onAuth({ url }) {
+        fillCalled = true
+        return {
+          username: 'testuser',
+          password: 'testpassword',
+        }
+      },
+      async onAuthSuccess(auth) {
+        approvedCalled = true
+      },
+      async onAuthFailure(auth) {
+        rejectedCalled = true
+      },
+    })
+    expect(fillCalled).toBe(true)
+    expect(approvedCalled).toBe(true)
+    expect(rejectedCalled).toBe(false)
+  })
+  it('onAuthFailure', async () => {
+    // Setup
+    const { fs, gitdir } = await makeFixture('test-push')
+    await setConfig({
+      fs,
+      gitdir,
+      path: 'remote.auth.url',
+      value: `http://${localhost}:8888/test-push-server-auth.git`,
+    })
+    // Test
+    let fillCalled = false
+    let approvedCalled = false
+    let rejectedCalled = false
+    let err
+    try {
+      await push({
+        fs,
+        http,
+        gitdir,
+        remote: 'auth',
+        ref: 'master',
+        async onAuth({ url }) {
+          fillCalled = true
+          return {
+            username: 'testuser',
+            password: 'NoT_rIgHt',
+          }
+        },
+        async onAuthSuccess(auth) {
+          approvedCalled = true
+        },
+        async onAuthFailure(auth) {
+          rejectedCalled = true
+        },
+      })
+    } catch (e) {
+      err = e
+    }
+    expect(err).toBeDefined()
+    expect(err.code).toBe(E.HTTPError)
+    expect(fillCalled).toBe(true)
+    expect(approvedCalled).toBe(false)
+    expect(rejectedCalled).toBe(true)
   })
 })
